Install latest git in devcontainer
This brings in a version of git that supports e.g. ssh commit signing. Since the host machine's .gitconfig gets copied into the container, we need a version of git that supports the options in that config. The default git from debian was too old to support SSH commit signing, which caused `task build` to fail, because it runs `go build`, which in turn runs a git command, which finally barfs with a config option it does not recognize.

Actually making ssh commit signing work in the devcontainer is not part of this.