Implement saving rendered recipes to single file
Put everything as multiple YAML documents in one file. The Recipe.Save API pretending that we're saving a single recipe is a bit wonky, maybe some sort of "Project" abstraction would be nice? It would encapsulate a bunch of recipes and it would be saved and loaded as a whole.