Change upgrade test to fail unless checksum checked
Modify a file in the new recipe version, which should go through as an
upgrade without user intervention, but it doesn't because we do not
yet see if rendered files still match their checksums. In this case
it matches and should be overwritten without asking (because it hasn't
been modified by the user).