Add Dockerfile for action
This is a multistage Dockerfile that builds the binary using task
and packages it up in a debian-based image with some OCI standard
labels applied.

The following labels should be applied by the builder (e.g. CI):
- org.opencontainers.image.created (RFC3339 datetime)
- org.opencontainers.image.version (version number)
- org.opencontainers.image.revision (VCS revision)

The taskfile version can be specified as a build argument along with
the sha256sum of the corresponding amd64 .deb package file.